requirements.txt
flask==3.0.3
google-generativeai==0.8.3
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
Pillow==10.4.0
gunicorn==22.0.0
